requirements.txt
# LaTF (LaRE + TruFor Fusion) - Python Dependencies
# 安装命令: pip install -r requirements.txt
# 注意: 请先安装好PyTorch (保持你现有的版本)

# ============================================
# 核心依赖 (PyTorch生态) - 不自动安装，使用已有版本
# ============================================
# torch        # 请手动安装，保持现有版本
# torchvision  # 请手动安装，保持现有版本

# ============================================
# CLIP相关 (使用open_clip替代原始CLIP)
# ============================================
open-clip-torch
ftfy
regex

# OpenAI CLIP（本项目 model.py 使用 `import clip`）
# 若环境缺失可用如下方式安装（需要git）：
# git+https://github.com/openai/CLIP.git

# ============================================
# 扩散模型 (LaRE特征提取)
# ============================================
diffusers
transformers
accelerate

# ============================================
# 图像处理
# ============================================
opencv-python
Pillow
albumentations

# ============================================
# 科学计算与机器学习
# ============================================
numpy
scipy
scikit-learn

# ============================================
# ViT & CNN Backbones (V11)
# ============================================
timm
# 可视化（训练曲线出图）
# ============================================
matplotlib

# ============================================
# 其他工具
# ============================================
huggingface-hub
safetensors

# ============================================
# 配置读取
# ============================================
python-dotenv

